[Refactor] FloatingMuenu에서 아이템 분리
